+++ src-tauri/src/models.rs
@@ -3,33 +3,33 @@ use specta::Type;
 
 #[derive(Debug, Serialize, Clone, Type)]
 pub struct Artists {
-    pub id: i32,
+    pub id: u32,
     pub name: String,
     pub path: String,
 }
 
 #[derive(Debug, Serialize, Clone, Type)]
 pub struct Albums {
-    pub id: i32,
-    pub artists_id: i32,
+    pub id: u32,
+    pub artists_id: u32,
     pub artist: String,
     pub name: String,
     pub cover_path: String,
     pub album_type: String,
-    pub duration: i32,
-    pub track_count: i32,
-    pub year: i32,
+    pub duration: u32,
+    pub track_count: u32,
+    pub year: u16,
     pub path: String,
 }
 
 #[derive(Debug, Serialize, Clone, Type)]
 pub struct Tracks {
-    pub id: i32,
-    pub duration: i32,
+    pub id: u32,
+    pub duration: u32,
     pub album: String,
-    pub albums_id: i32,
+    pub albums_id: u32,
     pub artist: String,
-    pub artists_id: i32,
+    pub artists_id: u32,
     pub name: String,
     pub path: String,
     pub cover_path: String,
@@ -37,7 +37,7 @@ pub struct Tracks {
 
 #[derive(Debug, Serialize, Clone, Type)]
 pub struct Playlists {
-    pub id: i32,
+    pub id: u32,
     pub name: String,
     pub description: String,
     pub cover_path: String,
